requirements.txt
#python3.6
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
Django==2.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
httpie==2.0.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2

